Change shunting yard and rpn calculator to support variables
shunting yard and rpn calculator now both support variables.

for now, the repl will ask the user to enter a variable's value
each time it appears, but in the future, something like a hashmap
could be used to store variable values.

no new leaks should be introduced by this commit, and this commit
allows the previous commit to build.