add run_function ability to run arbitrary functions by name
so this is not integrated into the repl yet, but there is now the
ability to call any function that takes doubles as arguments and
returns a double. all that is needed is the name, # of args, and
the args themselves.

there is still a bit of hardening to do, especially if the function
does not actually take the correct # of args (goes into the stack
and screws things up), or if the function expects non-double
arguments and ends up taking args from `rdi`, `rsi`, etc...

the end goal for this is to allow random people to use leo on a
server i run... i feel like i'm like the only person who is
silly enough to allow execution of random functions...